% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_hybrid_chart.R
\name{build_hybrid_chart}
\alias{build_hybrid_chart}
\title{build hybrid chart}
\usage{
build_hybrid_chart(data, community, adjust = FALSE)
}
\arguments{
\item{data}{}

\item{community}{}

\item{adjust}{}
}
\value{

}
\description{
build hybrid chart
}
