% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_data.R
\name{final_data}
\alias{final_data}
\title{final_data}
\usage{
final_data(data_frame)
}
\arguments{
\item{data_frame}{}
}
\value{

}
\description{
create data frame for confirmed cases to be put into the phase determination function
}
