% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_phase_dates.R
\name{find_phase_dates}
\alias{find_phase_dates}
\title{Find phase dates}
\usage{
find_phase_dates(data, adjust, ghost = TRUE)
}
\arguments{
\item{data}{}

\item{adjust}{}

\item{ghost}{}
}
\value{

}
\description{
function to determine phases with control limits and center lines
}
