% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_function.R
\name{graph_function}
\alias{graph_function}
\title{Title}
\usage{
graph_function(
  data,
  adjust = FALSE,
  type = if_else(adjust, "Adjusted", "Raw"),
  var_name = "Confirmed Cases"
)
}
\arguments{
\item{data}{}
}
\value{

}
\description{
function to graph
}
