% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_phase_change.R
\name{model_phase_change}
\alias{model_phase_change}
\title{Model phase changes
function to identify changes in phase}
\usage{
model_phase_change(data, y)
}
\arguments{
\item{data}{}

\item{y}{}
}
\value{

}
\description{
Model phase changes
function to identify changes in phase
}
